+++ src/main/java/com/example/demo/model/dto/request/ElectricityPayInfoRequest.java
@@ -1,12 +1,12 @@
 package com.example.demo.model.dto.request;
 
-import com.example.demo.model.db.entity.Plot;
 import com.example.demo.model.enums.PayStatus;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -18,7 +18,7 @@ import java.time.LocalDateTime;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ElectricityPayInfoRequest {
     Long plotNo;
-    Long amount;
+    BigDecimal amount;
     LocalDateTime date;
     PayStatus status;
 }
