+++ src/main/java/com/example/demo/model/dto/request/PlotPayInfoRequest.java
@@ -1,12 +1,12 @@
 package com.example.demo.model.dto.request;
 
-import com.example.demo.model.db.entity.Plot;
 import com.example.demo.model.enums.PayStatus;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -20,7 +20,7 @@ public class PlotPayInfoRequest {
     Long plotNo;
     String payDescription;
     LocalDateTime expectedDate;
-    Long amount;
+    BigDecimal amount;
     LocalDateTime factDate;
     PayStatus status;
 }
